[FEATURE]: moved map state and logic to hoc
