+++ src/components/map/map.test.js
@@ -1,30 +1,20 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import Map from './map.jsx';
+import {Map} from './map.jsx';
 
 const mock = {
-  coords: [
-    [52.3909553943508, 4.85309666406198],
-    [52.369553943508, 4.85309666406198],
-    [52.3909553943508, 4.929309666406198],
-    [52.3809553943508, 4.939309666406198]
-  ],
-  className: `test__map`
+  className: `test__map`,
+  mapRef: React.createRef()
 };
 
 it(`<Map/> rendered correctly`, () => {
-  const {coords, className} = mock;
-  const mockedHtmlElement = document.createElement(`div`);
+  const {className, mapRef} = mock;
 
   const tree = renderer
     .create(<Map
-      coords={coords}
       className={className}
-    />, {
-      createNodeMock: () => {
-        return mockedHtmlElement;
-      }
-    })
+      mapRef={mapRef}
+    />)
     .toJSON();
 
   expect(tree).toMatchSnapshot();
