+++ src/components/offer-list/offer-list.jsx
@@ -1,32 +1,19 @@
-import React, {PureComponent} from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 
-class OfferList extends PureComponent {
-  render() {
-    const {
-      offers,
-      onOfferHeaderClick,
-      className,
-      offerComponent: Offer,
-      onOfferCardMouseEnter
-    } = this.props;
-
-    return (
-      <div className={className}>
-        {offers.map((offer) =>
-          <Offer
-            key={offer.id}
-            offer={offer}
-            onOfferHeaderClick={onOfferHeaderClick}
-            onMouseEnter={(card) => {
-              this.setState({activeCard: card});
-              onOfferCardMouseEnter(offer);
-            }}
-          />)}
-      </div>
-    );
-  }
-}
+const OfferList = ({offers, onOfferHeaderClick, className, offerComponent: Offer, onOfferCardMouseEnter}) => {
+  return (
+    <div className={className}>
+      {offers.map((offer) =>
+        <Offer
+          key={offer.id}
+          offer={offer}
+          onOfferHeaderClick={onOfferHeaderClick}
+          onMouseEnter={onOfferCardMouseEnter}
+        />)}
+    </div>
+  );
+};
 
 OfferList.propTypes = {
   offerComponent: PropTypes.func.isRequired,
